Update ambari-bootstrap.sh
I added the download of Oracle Java RPM  version jdk-7u75-linux-x64.rpm.    This is a personal preference.    I also add the process to download the UnlimitedJCEPolicyJDK7.zip  this needed when you decided to implement  Secure cluster (Kerberos) within the cluster.  I still need to unzip and copy the jar to the proper location.   I also create a script "java.sh"   to /etc/profile.d/   this will set the JAVA_HOME and put in the PATH.

I also added